+++ CHANGELOG.md
@@ -14,7 +14,12 @@
     - Rename `ResetPos` to `ResetPosition`
     - Rename `SavePos` to `SavePosition`
     - Remove re-export cursor module types at root level, are now accessible from `crossterm::cursor`
-     
+ - `style module`
+    - Rename `ObjectStyle` to `ContentStyle`. Now full names are used for methods.
+    - Rename `StyledObject` to `StyledContent` and made members private. 
+    - Rename `attr` method to `attribute`.
+    - Rename `Attribute::NoInverse` to `NoReverse`
+  
 # Version 0.12.1
 
 - All the `crossterm_` crates code was moved to the `crossterm` crate